fix(vm): sync.go, snapshot.go, stats.go build tag 제거 및 상수 타입 변환
- sync.go, snapshot.go, stats.go: build tag 제거 (이미 driver 인터페이스 사용)
- snapshot_constants.go, stats_constants.go: uint32로 명시적 타입 변환

이제 CI-safe 빌드와 libvirt 빌드 모두 성공

+++ backend/internal/vm/snapshot_constants.go
@@ -7,10 +7,10 @@ import libvirt "github.com/libvirt/libvirt-go"
 
 // Snapshot constants from libvirt
 var (
-	SnapshotCreateAtomic      uint32 = libvirt.DOMAIN_SNAPSHOT_CREATE_ATOMIC
-	SnapshotCreateDiskOnly    uint32 = libvirt.DOMAIN_SNAPSHOT_CREATE_DISK_ONLY
-	SnapshotDeleteChildren    uint32 = libvirt.DOMAIN_SNAPSHOT_DELETE_CHILDREN
-	SnapshotRevertRunning     uint32 = libvirt.DOMAIN_SNAPSHOT_REVERT_RUNNING
-	SnapshotRevertForce       uint32 = libvirt.DOMAIN_SNAPSHOT_REVERT_FORCE
-	SnapshotDeleteMetadataOnly uint32 = libvirt.DOMAIN_SNAPSHOT_DELETE_METADATA_ONLY
+	SnapshotCreateAtomic      uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_CREATE_ATOMIC)
+	SnapshotCreateDiskOnly    uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_CREATE_DISK_ONLY)
+	SnapshotDeleteChildren    uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_DELETE_CHILDREN)
+	SnapshotRevertRunning     uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_REVERT_RUNNING)
+	SnapshotRevertForce       uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_REVERT_FORCE)
+	SnapshotDeleteMetadataOnly uint32 = uint32(libvirt.DOMAIN_SNAPSHOT_DELETE_METADATA_ONLY)
 )
